requirements.txt
streamlit==1.28.0
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
matplotlib==3.7.1
opencv-python==4.8.0.76
pillow==9.5.0